(stickies) add requested erroraction

+++ automatic/stickies/tools/chocolateybeforemodify.ps1
@@ -1,3 +1,5 @@
-﻿if (Get-Process $env:ChocolateyPackageName -ErrorAction SilentlyContinue) {
+﻿$ErrorActionPreference = 'Stop'
+
+if (Get-Process $env:ChocolateyPackageName -ErrorAction SilentlyContinue) {
    Stop-Process -Name $env:ChocolateyPackageName -Force
 }
